old-tests/webapps/IndexedDB/tests/submissions/Microsoft/idbobjectstore_deleteIndex.htm

<!DOCTYPE html>
<html>
<head>
    <meta content="text/html; charset=utf-8" http-equiv="content-type" />
    <title> IDBObjectStore.deleteIndex() - removes the index </title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript">
        var objectStoreName = "objectstore";
        var indexName = "i1"
        var db = null;
        
        var t = async_test();
        
        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName, databaseVersion);
            rqOpen.onupgradeneeded = t.step_func(function(event) {
                db = event.target.result;
                db.onerror = t.step_func(assert_database_error);
                var objStore = db.createObjectStore(objectStoreName);
                objStore.createIndex(indexName, "indexedProperty");
            });
            rqOpen.onerror = t.step_func(assert_open_request_error);
            rqOpen.onsuccess = t.step_func(function(event) {
                db.close();
                db = null;
                DeleteTheIndex();
            });
        }
        
        function DeleteTheIndex() {
            databaseVersion++;
            var rqOpen = window.indexedDB.open(databaseName, databaseVersion);
            rqOpen.onupgradeneeded = t.step_func(function(event) {
                db = event.target.result;
                db.onerror = t.step_func(assert_database_error);
                var objStore = event.target.transaction.objectStore(objectStoreName);
                objStore.deleteIndex(indexName);
            });
            rqOpen.onerror = t.step_func(assert_open_request_error);
            rqOpen.onsuccess = t.step_func(VerifyIndexDoesNotExist);
        }
        
        function VerifyIndexDoesNotExist() {
            try {
                var txn = db.transaction(objectStoreName, "readonly");
                var objStore = txn.objectStore(objectStoreName);
                var index = objStore.index(indexName);
                
                assert_expected_exception();
            } catch (ex) {
                assert_equals(ex.name, "NotFoundError", 'ex.name');
                t.done();
            }
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
